+++ src/lib.rs
@@ -31,6 +31,7 @@ pub mod mpv;
 pub mod mpv_render;
 pub mod mxf_wrap;
 pub mod otioz_import;
+pub mod package_edit;
 pub mod packaging;
 pub mod picture_processing;
 pub mod pipeline;